@@ -10,7 +10,7 @@ namespace Billing.DTO
 {
     public class SpecialisationDto : ProductTypeDto
     {
-        public SpecialisationDto(Specialisation specialisation, bool main) : base(specialisation.ProductType, false)
+        public SpecialisationDto(Specialisation specialisation, bool main, decimal ratio = 0) : base(specialisation.ProductType, false)
         {
             if (specialisation == null)
                 return;
@@ -22,10 +22,11 @@ namespace Billing.DTO
             this.SpecialisationName = specialisation.Name;
             this.SpecialisationId = specialisation.Id;
             this.DiscountType = specialisation.ProductType.DiscountType;
+            this.Ratio = ratio;
         }
         [DisplayName("Название специализации")]
         public int SpecialisationId { get; set; }
         public string SpecialisationName { get; set; }
-
+        public decimal Ratio { get; set; }
     }
 }
